move monaco them files to config
