+++ src/config/monaco-themes/atlantic-night.json.d.ts
